requirements.txt
# 核心依赖
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# 机器学习和NLP
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# API和HTTP
requests>=2.31.0
openai>=1.0.0

# 命令行和配置
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
tqdm>=4.65.0

# 数据处理和分析
xlsxwriter>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发和测试
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# 可选：本地LLM支持
# llama-cpp-python>=0.2.0
# ctransformers>=0.2.0
